fix: address iOS code review findings
- Fix weekday header ForEach collision by using enumerated indices
- Replace hardcoded 2026 fallback year with dynamic Calendar.current
- Add CodingKeys for snake_case JSON mapping in WeightStatsResponse,
  DayProperties, and DayPropertiesSet
- Throw error instead of fabricating DayProperties with empty userId
- Fix tab selection label to say "up to 3 tabs"
- Remove empty Text placeholders in TabSelectionView
- Restore NOVA group default to localized "Unknown"
- Extract MacroRadarView to its own file
- Restore dynamic target date in copy confirmation dialog
