+++ mobile/iosApp/Bissbilanz/Models/DayProperties.swift
@@ -4,6 +4,12 @@ struct DayProperties: Codable {
     let date: String
     let userId: String
     let isFastingDay: Bool
+
+    enum CodingKeys: String, CodingKey {
+        case date
+        case userId = "user_id"
+        case isFastingDay = "is_fasting_day"
+    }
 }
 
 struct DayPropertiesResponse: Codable {
@@ -12,4 +18,8 @@ struct DayPropertiesResponse: Codable {
 
 struct DayPropertiesSet: Codable {
     let isFastingDay: Bool
+
+    enum CodingKeys: String, CodingKey {
+        case isFastingDay = "is_fasting_day"
+    }
 }
